+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 httsleep Changelog
 ==================
 
+Version NEXT
+-------------
+* The kwarg ``verify`` is now supported, allowing users of httsleep to specify
+  ``verify=False`` in the same way as when directly using the ``requests`` library.
+
 Version 0.2.0
 -------------
 * The shorthand kwargs (``status_code``, ``json``, ``jsonpath``, ``text``, ``callback``)